Communication to Docusign Oauth is good but implementation needs work
I am able to get the DocuSign oauth working but only when redirecting to DocuSign's authentication page from a static page so I need to set a static page up for us to connect to. I am currently unable to get the static page to load when running the application.