Redirecting works now but I am unable to get information on the recipients of envelopes
After a while I was able to get the Oauth token from the DocuSign objects but I am unable to normally call any RESTful services using that token for some reason. Things break down and I have tested to make sure that the token is the actual authorization token used for the calls that are made to get the envelope information I do get.